my head hurts
→  [B-23]
→  Add the main menu system [#0]
→  Add a low graphics setting
→  Simplify and expand the music system
→  Normalize volume levels of all sounds and music
→  Fix NPCs not playing sounds when far from their original position

+++ scripts/scr_menu_pse/scr_menu_pse.gml
@@ -5,11 +5,14 @@ function fn_menu_pse_resume() // resumes the game
 	lvlTrans_tgtLvl = lvl_amtMax;
 	lvlTrans_kill = true;
 	
-	if (file_exists(scrSv_fname) == true)
-		file_delete(scrSv_fname);
-	if (sprite_exists(scrSv_spr) == true)
-		sprite_delete(scrSv_spr);
-	scrSv_spr = -1;
+	if (scrSv_spr != -1)
+	{
+		if (file_exists(scrSv_fname) == true)
+			file_delete(scrSv_fname);
+		if (sprite_exists(scrSv_spr) == true)
+			sprite_delete(scrSv_spr);
+		scrSv_spr = -1;
+	}
 	
 	instance_activate_all();
 	audio_resume_all();
